requirements.txt
anyio==4.12.1
APScheduler==3.11.2
certifi==2026.2.25
charset-normalizer==3.4.5
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
python-telegram-bot==22.6
requests==2.32.5
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
